brand-engine/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "brand-engine"
version = "0.1.0"
description = "Brand compliance engine: Gemini Embed 2 + Cohere v4 dual-fusion retrieval, drift detection, and grading"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn==0.34.2",
    "google-genai==1.14.0",
    "cohere==5.13.8",
    "pinecone==6.0.2",
    "pydantic==2.11.1",
    "Pillow==11.2.1",
    "numpy==2.2.4",
    "supabase==2.13.0",
    "python-dotenv==1.1.0",
    "typer==0.15.2",
    "httpx==0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.5",
    "pytest-asyncio==0.26.0",
]

[project.scripts]
brand-engine = "brand_engine.cli.main:app"
brand-engine-api = "brand_engine.api.server:start"

[tool.hatch.build.targets.wheel]
packages = ["brand_engine"]

[tool.hatch.build.targets.sdist]
include = ["brand_engine/", "data/"]
